+++ src/main/java/FacturarFruta.java
@@ -1,8 +1,8 @@
 public class FacturarFruta implements Facturador {
 
-    private double IVA=0.19;
+    private float IVA=0.19f;
     @Override
-    public double calcularPrecioConIVA(short precio) {
+    public float calcularPrecioConIVA(short precio) {
         return precio*IVA+precio;
     }
 }
